# contracts/openapi/ordin-api-v1.json
{
  "components": {
    "schemas": {
      "AuthSessionResponse": {
        "additionalProperties": false,
        "properties": {
          "tokens": {
            "$ref": "#/components/schemas/TokenPairResponse"
          },
          "user": {
            "$ref": "#/components/schemas/UserResponse"
          }
        },
        "required": [
          "tokens",
          "user"
        ],
        "title": "AuthSessionResponse",
        "type": "object"
      },
      "ClientPlatform": {
        "enum": [
          "android",
          "ios",
          "windows",
          "macos"
        ],
        "title": "ClientPlatform",
        "type": "string"
      },
      "DeviceInput": {
        "additionalProperties": false,
        "properties": {
          "appVersion": {
            "maxLength": 32,
            "minLength": 1,
            "title": "Appversion",
            "type": "string"
          },
          "installationId": {
            "format": "uuid",
            "title": "Installationid",
            "type": "string"
          },
          "platform": {
            "$ref": "#/components/schemas/ClientPlatform"
          }
        },
        "required": [
          "installationId",
          "platform",
          "appVersion"
        ],
        "title": "DeviceInput",
        "type": "object"
      },
      "FieldProblem": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "field",
          "code",
          "message"
        ],
        "title": "FieldProblem",
        "type": "object"
      },
      "GoalType": {
        "enum": [
          "loseFat",
          "gainMuscle",
          "maintain",
          "healthyEating"
        ],
        "title": "GoalType",
        "type": "string"
      },
      "HealthProfileInputModel": {
        "additionalProperties": false,
        "properties": {
          "birthDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "currentWeightKg": {
            "anyOf": [
              {
                "pattern": "^(?:0\\.(?:0[1-9]|[1-9][0-9]?)|[1-9][0-9]{0,3}(?:\\.[0-9]{1,2})?)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentweightkg"
          },
          "expectedVersion": {
            "minimum": 0.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "goalType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalType"
              },
              {
                "type": "null"
              }
            ]
          },
          "heightCm": {
            "anyOf": [
              {
                "pattern": "^(?:0\\.(?:0[1-9]|[1-9][0-9]?)|[1-9][0-9]{0,2}(?:\\.[0-9]{1,2})?)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heightcm"
          },
          "targetWeightKg": {
            "anyOf": [
              {
                "pattern": "^(?:0\\.(?:0[1-9]|[1-9][0-9]?)|[1-9][0-9]{0,3}(?:\\.[0-9]{1,2})?)$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetweightkg"
          }
        },
        "required": [
          "expectedVersion"
        ],
        "title": "HealthProfileInputModel",
        "type": "object"
      },
      "HealthProfileResponse": {
        "additionalProperties": false,
        "properties": {
          "birthDate": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthdate"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "currentWeightKg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currentweightkg"
          },
          "dailyEnergyTargetKcal": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dailyenergytargetkcal"
          },
          "goalType": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GoalType"
              },
              {
                "type": "null"
              }
            ]
          },
          "heightCm": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heightcm"
          },
          "targetWeightKg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targetweightkg"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "userId": {
            "format": "uuid",
            "title": "Userid",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "userId",
          "birthDate",
          "heightCm",
          "currentWeightKg",
          "targetWeightKg",
          "goalType",
          "dailyEnergyTargetKcal",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "title": "HealthProfileResponse",
        "type": "object"
      },
      "HealthResponse": {
        "properties": {
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "HealthResponse",
        "type": "object"
      },
      "OtpChallengeInput": {
        "additionalProperties": false,
        "properties": {
          "deviceInstallationId": {
            "format": "uuid",
            "title": "Deviceinstallationid",
            "type": "string"
          },
          "phoneNumber": {
            "pattern": "^\\+[1-9][0-9]{7,14}$",
            "title": "Phonenumber",
            "type": "string"
          }
        },
        "required": [
          "phoneNumber",
          "deviceInstallationId"
        ],
        "title": "OtpChallengeInput",
        "type": "object"
      },
      "OtpChallengeResponse": {
        "additionalProperties": false,
        "properties": {
          "challengeId": {
            "format": "uuid",
            "title": "Challengeid",
            "type": "string"
          },
          "expiresAt": {
            "format": "date-time",
            "title": "Expiresat",
            "type": "string"
          },
          "resendAfterSeconds": {
            "title": "Resendafterseconds",
            "type": "integer"
          }
        },
        "required": [
          "challengeId",
          "expiresAt",
          "resendAfterSeconds"
        ],
        "title": "OtpChallengeResponse",
        "type": "object"
      },
      "OtpVerificationInput": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "pattern": "^[0-9]{6}$",
            "title": "Code",
            "type": "string"
          },
          "device": {
            "$ref": "#/components/schemas/DeviceInput"
          }
        },
        "required": [
          "code",
          "device"
        ],
        "title": "OtpVerificationInput",
        "type": "object"
      },
      "ProblemDetails": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "fieldErrors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldProblem"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fielderrors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "code",
          "traceId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "ReadinessResponse": {
        "properties": {
          "status": {
            "const": "ready",
            "default": "ready",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "ReadinessResponse",
        "type": "object"
      },
      "RefreshTokenInput": {
        "additionalProperties": false,
        "properties": {
          "deviceInstallationId": {
            "format": "uuid",
            "title": "Deviceinstallationid",
            "type": "string"
          },
          "refreshToken": {
            "maxLength": 512,
            "minLength": 32,
            "title": "Refreshtoken",
            "type": "string"
          }
        },
        "required": [
          "refreshToken",
          "deviceInstallationId"
        ],
        "title": "RefreshTokenInput",
        "type": "object"
      },
      "TokenPairResponse": {
        "additionalProperties": false,
        "properties": {
          "accessToken": {
            "title": "Accesstoken",
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "format": "date-time",
            "title": "Accesstokenexpiresat",
            "type": "string"
          },
          "refreshToken": {
            "title": "Refreshtoken",
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "format": "date-time",
            "title": "Refreshtokenexpiresat",
            "type": "string"
          },
          "tokenType": {
            "default": "Bearer",
            "title": "Tokentype",
            "type": "string"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "refreshToken",
          "refreshTokenExpiresAt"
        ],
        "title": "TokenPairResponse",
        "type": "object"
      },
      "UserPatchInput": {
        "additionalProperties": false,
        "properties": {
          "expectedVersion": {
            "minimum": 1.0,
            "title": "Expectedversion",
            "type": "integer"
          },
          "nickname": {
            "maxLength": 40,
            "minLength": 1,
            "title": "Nickname",
            "type": "string"
          }
        },
        "required": [
          "expectedVersion",
          "nickname"
        ],
        "title": "UserPatchInput",
        "type": "object"
      },
      "UserResponse": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "nickname": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nickname"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "nickname",
          "status",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "title": "UserResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "AccessToken": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Ordin API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/otp/challenges": {
      "post": {
        "operationId": "createOtpChallenge",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 8,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpChallengeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpChallengeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Request a sign-in code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/otp/challenges/{challengeId}/verify": {
      "post": {
        "operationId": "verifyOtpChallenge",
        "parameters": [
          {
            "in": "path",
            "name": "challengeId",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Challengeid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerificationInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Verify a sign-in code",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/sessions/current": {
      "delete": {
        "operationId": "deleteCurrentSession",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Sign out the current session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/auth/token/refresh": {
      "post": {
        "operationId": "refreshAuthToken",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenPairResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "summary": "Rotate a refresh token",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v1/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Check API health",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/ready": {
      "get": {
        "operationId": "getReadiness",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "503": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Check dependency readiness",
        "tags": [
          "system"
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get the current user",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "updateCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserPatchInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Update the current user",
        "tags": [
          "users"
        ]
      }
    },
    "/api/v1/users/me/health-profile": {
      "get": {
        "operationId": "getCurrentHealthProfile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Get the current health profile",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "putCurrentHealthProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HealthProfileInputModel"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unauthorized"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable Content"
          }
        },
        "security": [
          {
            "AccessToken": []
          }
        ],
        "summary": "Create or replace the current health profile",
        "tags": [
          "users"
        ]
      }
    }
  }
}